better fix tests by moving things
