+++ integration-tests/external-plugin-test/external-plugin/src/main/java/co/elastic/apm/agent/plugin/PluginInstrumentation.java
@@ -22,7 +22,7 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.plugin;
+package co.elastic.apm.agent.plugin;
 
 import co.elastic.apm.agent.sdk.ElasticApmInstrumentation;
 import co.elastic.apm.api.ElasticApm;
@@ -43,7 +43,7 @@ public class PluginInstrumentation extends ElasticApmInstrumentation {
 
     @Override
     public ElementMatcher<? super TypeDescription> getTypeMatcher() {
-        return named(System.getProperty("elastic.apm.plugin.instrumented_class", "co.elastic.apm.plugin.test.TestClass"));
+        return named(System.getProperty("elastic.apm.plugin.instrumented_class", "co.elastic.apm.agent.plugin.test.TestClass"));
     }
 
     @Override
