+++ integration-tests/external-plugin-test/external-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,1 +1,1 @@
-co.elastic.apm.plugin.PluginInstrumentation
+co.elastic.apm.agent.plugin.PluginInstrumentation
